+++ packages/mobile/src/tabs/Tabs.tsx
@@ -22,6 +22,9 @@ import { useComponentConfig } from '../hooks/useComponentConfig';
 import type { BoxBaseProps, BoxProps, HStackProps } from '../layout';
 import { Box, HStack } from '../layout';
 
+import { DefaultTab } from './DefaultTab';
+import { DefaultTabsActiveIndicator } from './DefaultTabsActiveIndicator';
+
 const AnimatedBox = Animated.createAnimatedComponent(Box);
 
 type TabContainerProps = {
@@ -49,31 +52,44 @@ export type TabsActiveIndicatorProps = {
   activeTabRect: Rect;
 } & BoxProps;
 
-export type TabComponentProps<TabId extends string = string> = TabValue<TabId> & {
+export type TabComponentProps<
+  TabId extends string = string,
+  TTab extends TabValue<TabId> = TabValue<TabId>,
+> = Omit<TTab, 'Component' | 'id'> & {
+  id: TabId;
   style?: StyleProp<ViewStyle>;
 };
 
-export type TabComponent<TabId extends string = string> = React.FC<TabComponentProps<TabId>>;
+export type TabComponent<
+  TabId extends string = string,
+  TTab extends TabValue<TabId> = TabValue<TabId>,
+> = React.FC<TabComponentProps<TabId, TTab>>;
 
 export type TabsActiveIndicatorComponent = React.FC<TabsActiveIndicatorProps>;
 
-export type TabsBaseProps<TabId extends string = string> = Omit<BoxBaseProps, 'onChange'> &
-  Omit<TabsOptions<TabId>, 'tabs'> & {
+export type TabsBaseProps<
+  TabId extends string = string,
+  TTab extends TabValue<TabId> = TabValue<TabId>,
+> = Omit<BoxBaseProps, 'onChange'> &
+  Omit<TabsOptions<TabId, TTab>, 'tabs'> & {
     /** The array of tabs data. Each tab may optionally define a custom Component to render. */
-    tabs: (TabValue<TabId> & { Component?: TabComponent<TabId> })[];
+    tabs: (TTab & { Component?: TabComponent<TabId, TTab> })[];
     /** The default Component to render each tab. */
-    TabComponent: TabComponent<TabId>;
+    TabComponent?: TabComponent<TabId, TTab>;
     /** The default Component to render the tabs active indicator. */
-    TabsActiveIndicatorComponent: TabsActiveIndicatorComponent;
+    TabsActiveIndicatorComponent?: TabsActiveIndicatorComponent;
     /** Background color passed to the TabsActiveIndicatorComponent. */
     activeBackground?: ThemeVars.Color;
     /** Optional callback to receive the active tab element. */
     onActiveTabElementChange?: (element: View | null) => void;
-    /** Custom styles for individual elements of the Tabs component */
   };
 
-export type TabsProps<TabId extends string = string> = TabsBaseProps<TabId> &
+export type TabsProps<
+  TabId extends string = string,
+  TTab extends TabValue<TabId> = TabValue<TabId>,
+> = TabsBaseProps<TabId, TTab> &
   Omit<HStackProps, 'onChange'> & {
+    /** Custom styles for individual elements of the Tabs component */
     styles?: {
       /** Root container element */
       root?: StyleProp<ViewStyle>;
@@ -84,106 +100,114 @@ export type TabsProps<TabId extends string = string> = TabsBaseProps<TabId> &
     };
   };
 
-type TabsFC = <TabId extends string = string>(
-  props: TabsProps<TabId> & { ref?: React.ForwardedRef<View> },
+type TabsFC = <TabId extends string = string, TTab extends TabValue<TabId> = TabValue<TabId>>(
+  props: TabsProps<TabId, TTab> & { ref?: React.ForwardedRef<View> },
 ) => React.ReactElement;
 
 const TabsComponent = memo(
-  forwardRef(<TabId extends string>(_props: TabsProps<TabId>, ref: React.ForwardedRef<View>) => {
-    const mergedProps = useComponentConfig('Tabs', _props);
-    const {
-      tabs,
-      TabComponent,
-      TabsActiveIndicatorComponent,
-      activeBackground,
-      activeTab,
-      disabled,
-      onChange,
-      styles,
-      style,
-      role = 'tablist',
-      position = 'relative',
-      alignSelf = 'flex-start',
-      opacity,
-      onActiveTabElementChange,
-      borderRadius,
-      borderTopLeftRadius,
-      borderTopRightRadius,
-      borderBottomLeftRadius,
-      borderBottomRightRadius,
-      ...props
-    } = mergedProps;
-    const tabsContainerRef = useRef<View>(null);
-    useImperativeHandle(ref, () => tabsContainerRef.current as View, []); // merge internal ref to forwarded ref
+  forwardRef(
+    <TabId extends string, TTab extends TabValue<TabId> = TabValue<TabId>>(
+      _props: TabsProps<TabId, TTab>,
+      ref: React.ForwardedRef<View>,
+    ) => {
+      const mergedProps = useComponentConfig('Tabs', _props);
+      const {
+        tabs,
+        TabComponent = DefaultTab,
+        TabsActiveIndicatorComponent = DefaultTabsActiveIndicator,
+        activeBackground,
+        activeTab,
+        disabled,
+        onChange,
+        styles,
+        style,
+        role = 'tablist',
+        position = 'relative',
+        alignSelf = 'flex-start',
+        opacity,
+        onActiveTabElementChange,
+        borderRadius,
+        borderTopLeftRadius,
+        borderTopRightRadius,
+        borderBottomLeftRadius,
+        borderBottomRightRadius,
+        testID,
+        ...props
+      } = mergedProps;
+      const tabsContainerRef = useRef<View>(null);
+      useImperativeHandle(ref, () => tabsContainerRef.current as View, []); // merge internal ref to forwarded ref
 
-    const refMap = useRefMap<View>();
-    const api = useTabs<TabId>({ tabs, activeTab, disabled, onChange });
+      const refMap = useRefMap<View>();
+      const api = useTabs<TabId, TTab>({ tabs, activeTab, disabled, onChange });
 
-    const [activeTabRect, setActiveTabRect] = useState<Rect>(defaultRect);
-    const previousActiveRef = useRef(activeTab);
+      const [activeTabRect, setActiveTabRect] = useState<Rect>(defaultRect);
+      const previousActiveRef = useRef(activeTab);
 
-    const updateActiveTabRect = useCallback(() => {
-      const activeTabRef = activeTab ? refMap.getRef(activeTab.id) : null;
-      if (!activeTabRef || !tabsContainerRef.current) return;
-      activeTabRef.measureLayout(tabsContainerRef.current, (x, y, width, height) =>
-        setActiveTabRect({ x, y, width, height }),
+      const updateActiveTabRect = useCallback(() => {
+        const activeTabRef = activeTab ? refMap.getRef(activeTab.id) : null;
+        if (!activeTabRef || !tabsContainerRef.current) return;
+        activeTabRef.measureLayout(tabsContainerRef.current, (x, y, width, height) =>
+          setActiveTabRect({ x, y, width, height }),
+        );
+      }, [activeTab, refMap]);
+
+      const registerRef = useCallback(
+        (tabId: string, ref: View) => {
+          refMap.registerRef(tabId, ref);
+          if (activeTab?.id === tabId) {
+            onActiveTabElementChange?.(ref);
+          }
+        },
+        [activeTab, onActiveTabElementChange, refMap],
       );
-    }, [activeTab, refMap]);
 
-    const registerRef = useCallback(
-      (tabId: string, ref: View) => {
-        refMap.registerRef(tabId, ref);
-        if (activeTab?.id === tabId) {
-          onActiveTabElementChange?.(ref);
-        }
-      },
-      [activeTab, onActiveTabElementChange, refMap],
-    );
+      if (previousActiveRef.current !== activeTab) {
+        previousActiveRef.current = activeTab;
+        updateActiveTabRect();
+      }
 
-    if (previousActiveRef.current !== activeTab) {
-      previousActiveRef.current = activeTab;
-      updateActiveTabRect();
-    }
-
-    return (
-      <HStack
-        ref={tabsContainerRef}
-        alignSelf={alignSelf}
-        borderBottomLeftRadius={borderBottomLeftRadius}
-        borderBottomRightRadius={borderBottomRightRadius}
-        borderRadius={borderRadius}
-        borderTopLeftRadius={borderTopLeftRadius}
-        borderTopRightRadius={borderTopRightRadius}
-        onLayout={updateActiveTabRect}
-        opacity={opacity ?? (disabled ? accessibleOpacityDisabled : 1)}
-        position={position}
-        role={role}
-        style={styles?.root ? [style, styles.root] : style}
-        {...props}
-      >
-        <TabsContext.Provider value={api as TabsApi<string>}>
-          <TabsActiveIndicatorComponent
-            activeTabRect={activeTabRect}
-            background={activeBackground}
-            borderBottomLeftRadius={borderBottomLeftRadius}
-            borderBottomRightRadius={borderBottomRightRadius}
-            borderRadius={borderRadius}
-            borderTopLeftRadius={borderTopLeftRadius}
-            borderTopRightRadius={borderTopRightRadius}
-            style={styles?.activeIndicator}
-          />
-          {tabs.map(({ id, Component: CustomTabComponent, disabled: tabDisabled, ...props }) => {
-            const RenderedTab = CustomTabComponent ?? TabComponent;
-            return (
-              <TabContainer key={id} id={id} registerRef={registerRef}>
-                <RenderedTab disabled={tabDisabled} id={id} style={styles?.tab} {...props} />
-              </TabContainer>
-            );
-          })}
-        </TabsContext.Provider>
-      </HStack>
-    );
-  }),
+      return (
+        <HStack
+          ref={tabsContainerRef}
+          alignSelf={alignSelf}
+          borderBottomLeftRadius={borderBottomLeftRadius}
+          borderBottomRightRadius={borderBottomRightRadius}
+          borderRadius={borderRadius}
+          borderTopLeftRadius={borderTopLeftRadius}
+          borderTopRightRadius={borderTopRightRadius}
+          onLayout={updateActiveTabRect}
+          opacity={opacity ?? (disabled ? accessibleOpacityDisabled : 1)}
+          position={position}
+          role={role}
+          style={styles?.root ? [style, styles.root] : style}
+          testID={testID}
+          {...props}
+        >
+          <TabsContext.Provider value={api as TabsApi<string>}>
+            <TabsActiveIndicatorComponent
+              activeTabRect={activeTabRect}
+              background={activeBackground}
+              borderBottomLeftRadius={borderBottomLeftRadius}
+              borderBottomRightRadius={borderBottomRightRadius}
+              borderRadius={borderRadius}
+              borderTopLeftRadius={borderTopLeftRadius}
+              borderTopRightRadius={borderTopRightRadius}
+              style={styles?.activeIndicator}
+              testID={testID ? `${testID}-active-indicator` : undefined}
+            />
+            {tabs.map(({ id, Component: CustomTabComponent, ...props }) => {
+              const RenderedTab = CustomTabComponent ?? TabComponent;
+              return (
+                <TabContainer key={id} id={id} registerRef={registerRef}>
+                  <RenderedTab id={id} style={styles?.tab} {...props} />
+                </TabContainer>
+              );
+            })}
+          </TabsContext.Provider>
+        </HStack>
+      );
+    },
+  ),
 );
 
 TabsComponent.displayName = 'Tabs';
@@ -194,6 +218,7 @@ export const TabsActiveIndicator = ({
   activeTabRect,
   position = 'absolute',
   style,
+  testID = 'tabs-active-indicator',
   ...props
 }: TabsActiveIndicatorProps) => {
   const previousActiveTabRect = useRef(activeTabRect);
@@ -224,7 +249,7 @@ export const TabsActiveIndicator = ({
       position={position}
       role="none"
       style={[animatedBoxStyle, style]}
-      testID="tabs-active-indicator"
+      testID={testID}
       {...props}
     />
   );